Moved old Table comp the TableDeprecated
