+++ frontend/src/lib/components/TableDeprecated/index.ts
@@ -1,0 +1,4 @@
+export {
+    /** @deprecated Use table/Table instead */
+    Table as TableDeprecated,
+} from "./table";
